refactor: cloud supported auth types
This commit moves a function from the cloud state for checking a cloud's
supported auth types and places this in the model controller state
logic. The implementation where it currently was didn't have any usages
and now needs to be used in model state where it can answer questions
around updating a model's cloud credential.
