+++ domain/cloud/state/state_test.go
@@ -8,7 +8,6 @@ import (
 	"database/sql"
 	"time"
 
-	"github.com/canonical/sqlair"
 	"github.com/juju/collections/set"
 	"github.com/juju/errors"
 	jc "github.com/juju/testing/checkers"
@@ -512,55 +511,6 @@ func (s *stateSuite) TestNullCloudType(c *gc.C) {
 		return err
 	})
 	c.Assert(jujudb.IsErrConstraintNotNull(err), jc.IsTrue)
-}
-
-// TestCloudSupportsAuthTypeTrue is asserting the happy path that for a valid
-// cloud and supported auth type we get back true with no errors.
-func (s *stateSuite) TestCloudSupportsAuthTypeTrue(c *gc.C) {
-	st := NewState(s.TxnRunnerFactory())
-	s.assertInsertCloud(c, st, testCloud)
-
-	var supports bool
-	err := s.TxnRunner().Txn(context.Background(), func(ctx context.Context, tx *sqlair.TX) error {
-		s, err := CloudSupportsAuthType(context.Background(), tx, testCloud.Name, testCloud.AuthTypes[0])
-		supports = s
-		return err
-	})
-
-	c.Assert(err, jc.ErrorIsNil)
-	c.Check(supports, jc.IsTrue)
-}
-
-// TestCloudSupportsAuthTypeFalse is asserting the happy path that for a valid
-// cloud and a non supported auth type we get back false with no errors.
-func (s *stateSuite) TestCloudSupportsAuthTypeFalse(c *gc.C) {
-	st := NewState(s.TxnRunnerFactory())
-	s.assertInsertCloud(c, st, testCloud)
-
-	var supports bool
-	err := s.TxnRunner().Txn(context.Background(), func(ctx context.Context, tx *sqlair.TX) error {
-		s, err := CloudSupportsAuthType(context.Background(), tx, testCloud.Name, cloud.AuthType("no-exist"))
-		supports = s
-		return err
-	})
-
-	c.Assert(err, jc.ErrorIsNil)
-	c.Check(supports, jc.IsFalse)
-}
-
-// TestCloudSupportsAuthTypeCloudNotFound is checking to that if we ask if a
-// cloud supports an auth type and the cloud doesn't exist we get back a
-// [clouderrors.NotFound] error.
-func (s *stateSuite) TestCloudSupportsAuthTypeCloudNotFound(c *gc.C) {
-	var supports bool
-	err := s.TxnRunner().Txn(context.Background(), func(ctx context.Context, tx *sqlair.TX) error {
-		s, err := CloudSupportsAuthType(context.Background(), tx, "no-exist", cloud.AuthType("no-exist"))
-		supports = s
-		return err
-	})
-
-	c.Assert(err, jc.ErrorIs, clouderrors.NotFound)
-	c.Check(supports, jc.IsFalse)
 }
 
 func (s *stateSuite) ensureUser(c *gc.C, userUUID, name, createdByUUID string) {
