+++ domain/model/state/types.go
@@ -18,6 +18,14 @@ import (
 	"github.com/juju/juju/core/user"
 )
 
+// cloudID represents only the name and UUID of a cloud
+type cloudID struct {
+	// UUID holds the cloud's unique identifier.
+	UUID string `db:"uuid"`
+	// Name holds the cloud name.
+	Name string `db:"name"`
+}
+
 // dbModel represents the state of a model.
 type dbModel struct {
 	// Name is the human friendly name of the model.
